// src/app/core/models/pedido.model.ts
export class Pedido {
  id: number;
  instituicao: number;
  descricao: string;
  numeroProtocolo: string;
  nome: string;
  data: Date;
  area: any;
  resposta: string;
  dataPrazo: Date;
  statusAtendimento: number;
  atendimentoDto: import("d:/Desenvolvimento/frontend/lai/src/app/core/models/atendimento.model").Atendimento;
  }