// src/app/core/models/recurso.model.ts
export class Recurso {
  id: number;
  atendimento: number;
  tipo: number;
  justificativa: string; 
  dataRecurso: Date;
  dataPrazo: Date;
  respostaRecurso: string;
  dataResposta: Date;
  status: number;
  descricaoTipo: string;
  mensagem: string;
  recursoAutorizado: any;
  descricaoStatus: String;
  }